Hard/25_Discord_Bot_requirements.txt
discord.py==2.3.2